pyproject.toml
[project]
name = "manifone-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.8",
    "librosa>=0.11.0",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "resemblyzer>=0.1.4",
    "scikit-learn>=1.7.1",
    "streamlit>=1.48.0",
    "torchaudio>=2.8.0",
    "transformers>=4.55.2",
    "umap>=0.1.1",
]
